Extract multiple comparison
